Remove image render API (merge with custom render API)
