+++ example/lib/main.dart
@@ -300,28 +300,23 @@ class _MyHomePageState extends State<MyHomePage> {
             svgDataUriMatcher(): svgDataImageRender,
             svgAssetUriMatcher(): svgAssetImageRender,
             svgNetworkSourceMatcher(): svgNetworkImageRender,
-            tableMatcher(): tableRender,
-            videoMatcher(): videoRender,
-          },
-          customImageRenders: {
-            networkSourceMatcher(domains: ["flutter.dev"]):
-                (context, attributes, element) {
+            networkSourceMatcher(domains: ["flutter.dev"]): CustomRender.fromWidget(
+                widget: (context, buildChildren) {
               return FlutterLogo(size: 36);
-            },
-            networkSourceMatcher(domains: ["mydomain.com"]):
-                networkImageRender(
+            }),
+            networkSourceMatcher(domains: ["mydomain.com"]): networkImageRender(
               headers: {"Custom-Header": "some-value"},
               altWidget: (alt) => Text(alt ?? ""),
               loadingWidget: () => Text("Loading..."),
             ),
             // On relative paths starting with /wiki, prefix with a base url
-            (attr, _) =>
-                    attr["src"] != null && attr["src"]!.startsWith("/wiki"):
-                networkImageRender(
-                    mapUrl: (url) => "https://upload.wikimedia.org" + url!),
+            (context) => context.tree.element?.attributes["src"] != null
+                && context.tree.element!.attributes["src"]!.startsWith("/wiki"):
+            networkImageRender(mapUrl: (url) => "https://upload.wikimedia.org" + url!),
             // Custom placeholder image for broken links
-            networkSourceMatcher():
-                networkImageRender(altWidget: (_) => FlutterLogo()),
+            networkSourceMatcher(): networkImageRender(altWidget: (_) => FlutterLogo()),
+            tableMatcher(): tableRender,
+            videoMatcher(): videoRender,
           },
           onLinkTap: (url, _, __, ___) {
             print("Opening $url...");
